+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 * Fixed an issue where in some occasions running of test-playbooks could receive session timeouts.
 * Fixed an issue where making a change in a module using an ApiModule would cause lint to run on the ApiModule unnecessarily.
-* Fixed an issue where the `marketplace` field was ignored where dumping pack zips.
+* Fixed an issue where the `marketplace` field was not used when dumping pack zips.
 
 ## 1.10.2
 * Added markdown lint formatting for README files in the **format** command.