+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 * Fixed an issue where in some occasions running of test-playbooks could receive session timeouts.
+* Fixed an issue where making a change in a module using an ApiModule would cause lint to run on the ApiModule unnecessarily.
 
 ## 1.10.2
 * Added markdown lint formatting for README files in the **format** command.