+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Unreleased
-* Fixed an issue where trying to run the mdx-server locally in the **format** command would fail.
+* Fixed an issue where trying to run node dependencies locally in the **format** command would fail.
 * Fixed an issue where in some occasions running of test-playbooks could receive session timeouts.
 * Fixed an issue where making a change in a module using an ApiModule would cause lint to run on the ApiModule unnecessarily.
 * Fixed an issue where the `marketplace` field was not used when dumping pack zips.