+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * Fixed an issue where in some occasions running of test-playbooks could receive session timeouts.
 * Fixed an issue where making a change in a module using an ApiModule would cause lint to run on the ApiModule unnecessarily.
 * Fixed an issue where the `marketplace` field was not used when dumping pack zips.
-* Updated integration generated README for new version of integrations.
+* Fixed a typo in the README content generated with **update-release-notes** for updating integrations.
 
 ## 1.10.2
 * Added markdown lint formatting for README files in the **format** command.